[feature] complete Section 25

+++ controller/feed.js
@@ -3,6 +3,8 @@ const fs = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
+const user = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -44,18 +46,28 @@ exports.createPost = (req, res, next) => {
   const title = req.body.title;
   const content = req.body.content;
   const imageUrl = req.file.path;
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: 'Maximilian' }
+    creator: req.userId
   });
   post
     .save()
     .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
         message: 'Post created successfully!',
-        post: result
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
       });
     })
     .catch(err => {
@@ -109,7 +121,12 @@ exports.updatePost = (req, res, next) => {
     if (!post) {
       const error = new Error('Could not Find Data');
       error.statusCode = 404;
-      throw error
+      throw error;
+    }
+    if (post.creator.toString() !== req.userId) {
+      const error = new Error('Not Authorized!!!');
+      error.statusCode = 403;
+      throw error;
     }
     if (imageUrl !== post.imageUrl) {
       clearImage(post.imageUrl)
@@ -140,11 +157,22 @@ exports.deletePost = (req, res, next) => {
       error.statusCode = 404;
       throw error;
     }
+    if (post.creator.toString() !== req.userId) {
+      const error = new Error('Not Authorized!!!');
+      error.statusCode = 403;
+      throw error;
+    }
     clearImage(post.imageUrl);
     return Post.findByIdAndRemove(postId);
   })
   .then(result => {
-    console.log(result);
+    return User.findById(req.userId)
+  })
+  .then(user => {
+    user.posts.pull(postId);
+    return user.save()
+  })
+  .then(result => {
     res.status(200).json({ message: 'Deleted Post.'})
   })
   .catch(err => {
